requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.18.4
aiofiles==24.1.0
python-multipart==0.0.22
pydantic==2.11.1
Pillow==12.1.1
fonttools==4.62.1
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
